// src/common/enum/common.enum.ts
export enum SortDirection {
  ASC = 'ASC',
  DESC = 'DESC',
}

export enum TriggerType {
  WS = 'WS',
}

export enum ExecutionLayer {
  BASE_CHAIN = 'BASE_CHAIN',
  EPHEMERAL_ROLLUP = 'EPHEMERAL_ROLLUP',
}

export enum VertexTransactionType {
  INIT_USER_VAULT = 'INIT_USER_VAULT',
  DELEGATE_USER_VAULT = 'DELEGATE_USER_VAULT',
  INIT_INDEXER = 'INIT_INDEXER',
  DEPOSIT_TO_VAULT = 'DEPOSIT_TO_VAULT',
  WITHDRAW_INDEXER_FEE = 'WITHDRAW_INDEXER_FEE',
  TRACK_USER_ACTIVITY = 'TRACK_USER_ACTIVITY',
  START_BILLING = 'START_BILLING',
  COMMIT_AND_START_BILLING = 'COMMIT_AND_START_BILLING',
  UNDELEGATED_AND_COMMIT_USER_VAULT = 'UNDELEGATED_AND_COMMIT_USER_VAULT',
  CHARGED_FEE = 'CHARGED_FEE',
}